Create yield_fixture for modified_httpretty_asserter
