+++ src/shipchain_common/test_utils/__init__.py
@@ -41,6 +41,7 @@ from .mocked_rpc_responses import \
     valid_eth_amount
 
 from .httpretty_asserter import \
-    HTTPrettyAsserter
+    HTTPrettyAsserter, \
+    modified_http_pretty
 
 from ..utils import validate_uuid4  # Imported for backwards compatibility usage of `from shipchain_common.test_utils`
