+++ src/shipchain_common/test_utils/httpretty_asserter.py
@@ -1,3 +1,19 @@
+"""
+Copyright 2020 ShipChain, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+import pytest
 from httpretty import HTTPretty
 
 from urllib.parse import urlparse
@@ -56,3 +72,10 @@ class HTTPrettyAsserter(HTTPretty):
         if 'query' in assertion:
             assert assertion['query'] == call['query'], \
                 f'Error: query mismatch, desired `{assertion["query"]}` returned `{call["query"]}`.'
+
+
+@pytest.yield_fixture
+def modified_http_pretty():
+    HTTPrettyAsserter.enable(allow_net_connect=False)
+    yield HTTPrettyAsserter
+    HTTPrettyAsserter.disable()
